pyproject.toml
[project]
name = "python-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "litellm>=1.72.0",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
